blk: Rename the ramhd_sq to ramhd_bio
Essentially, the bio-based driver is irrelevant with the
single-queue or multi-queue, the latter is some kinds of
request-based, spcifically the kernel block IO layer will
handle the raw bios with sq or mq. Given the sq is deprecated
in the latest kernel, so the only focus for us is the
blk-mq model. Forget the 'blk-sq' if any...

Signed-off-by: xuesong chen <xuesong.cxs@outlook.com>