blk: dump the partition table of the gendisk
All the partition devices(include the gendisk->part0 itself) are stored into
the gendisk->part_tbl xarray. With the help of the kprobe of
'add_partition(...), it's easy to dump this partition table infomation via
xa_load(...) method.

Please note that since the kprobe is located in the entrance of the
add_partition(...), so the current 'partno' will not be dumped because the
add_partition(...) has not added this 'partno' into gendisk->part_tbl yet.

Signed-off-by: xuesong chen <xuesong.cxs@outlook.com>